updata

+++ core/src/caliniya/armavoke/base/game/WorldChunk.java
@@ -10,9 +10,9 @@ public class WorldChunk {
   public static final int SHIFT = 5; 
 
   // 数据层 1: 地板
-  public short[] floorIds;
+  public int[] floorIds;
   // 数据层 2: 环境方块 (墙壁等)
-  public short[] envblockIds;
+  public int[] envblockIds;
   // 数据层 3: 建筑实例
   public Building[] buildings;
     //不存储单位
@@ -20,25 +20,25 @@ public class WorldChunk {
   public boolean empty = true;
 
   public WorldChunk() {
-    floorIds = new short[AREA];
-    envblockIds = new short[AREA];
+    floorIds = new int[AREA];
+    envblockIds = new int[AREA];
     buildings = new Building[AREA]; // 初始化引用数组
   }
   
   // --- Floor ---
-  public short getFloor(int localX, int localY) {
+  public int getFloor(int localX, int localY) {
     return floorIds[(localY << SHIFT) | localX];
   }
-  public void setFloor(int localX, int localY, short id) {
+  public void setFloor(int localX, int localY, int id) {
     floorIds[(localY << SHIFT) | localX] = id;
     empty = false;
   }
 
   // --- ENVBlock ---
-  public short getENVBlock(int localX, int localY) {
+  public int getENVBlock(int localX, int localY) {
     return envblockIds[(localY << SHIFT) | localX];
   }
-  public void setENVBlock(int localX, int localY, short id) {
+  public void setENVBlock(int localX, int localY, int id) {
     envblockIds[(localY << SHIFT) | localX] = id;
     empty = false;
   }
